+++ options/db_options.h
@@ -83,6 +83,7 @@ struct ImmutableDBOptions {
   bool avoid_unnecessary_blocking_io;
   bool persist_stats_to_disk;
   size_t log_readahead_size;
+  bool enable_db_mutex_owned_timer;
 };
 
 struct MutableDBOptions {